python-api/requirements.txt
kafka-python==2.0.2
python-dotenv==0.20.0
fastapi[all]==0.75.1